requirements.txt
aiohttp==3.12.13
fitz==0.0.1.dev2
llama_index==0.12.42
matplotlib==3.10.3
numpy==2.3.0
openai==1.88.0
pandas==2.3.0
PyYAML==6.0.2
rapidfuzz==3.13.0
Requests==2.32.4
scikit_learn==1.7.0
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
huggingface_hub==0.34.4
transformers
torch
llama-index-embeddings-huggingface
python-dotenv==1.0.1
docling
